% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Algorithm.R, R/methods-Algorithm.R
\docType{class}
\name{Algorithm-class}
\alias{Algorithm-class}
\alias{Algorithm}
\title{Wrapper function Algorithm}
\usage{
Algorithm(serverURL, name)
}
\arguments{
\item{serverURL}{a WTSCS server URL}

\item{name}{an Algorithm name}
}
\description{
Wrapper function Algorithm
}
\section{Slots }{

\describe{   
  \item{\code{name}:}{Attribute of class \code{"character"}, Name of the Algorithm.}
  \item{\code{input_parameters}:}{Attribute of class \code{"character"}, Input parameters of the Algorithm}
  \item{\code{output}:}{Attribute of class \code{"character"}, Output columns of the Algorithm}
  \item{\code{description}:}{Attribute of class \code{"character"}, Description of the Algorithm}
}
}

