% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Algorithm.R
\docType{methods}
\name{applyAlgorithm}
\alias{applyAlgorithm}
\alias{applyAlgorithm-generic}
\alias{applyAlgorithm,Algorithm,character-method}
\title{Assign Algorithm parameters to an Algorithm object}
\usage{
applyAlgorithm(alg, algParameters)

\S4method{applyAlgorithm}{Algorithm,character}(alg, algParameters)
}
\arguments{
\item{alg}{an Algorithm object}

\item{algParameters}{Algorithm parameters}
}
\description{
Assign Algorithm parameters to an Algorithm object
}
