% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-WTSPS.R
\docType{methods}
\name{cancelProcess}
\alias{cancelProcess}
\alias{cancelProcess-generic}
\alias{cancelProcess,ANY-method}
\title{Returns a Process status in a WTSPS server URL}
\usage{
cancelProcess(serverInfo, processInfo)

\S4method{cancelProcess}{ANY}(serverInfo, processInfo)
}
\arguments{
\item{serverInfo}{a WTSPS object or URL}

\item{processInfo}{a Process object}
}
\description{
Returns a Process status in a WTSPS server URL
}
