% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Algorithm.R
\docType{methods}
\name{describeAlgorithm}
\alias{describeAlgorithm}
\alias{describeAlgorithm-generic}
\alias{describeAlgorithm,ANY,character-method}
\title{Returns an Algorithm Class in a WTSPS server URL queried by name}
\usage{
describeAlgorithm(serverInfo, name)

\S4method{describeAlgorithm}{ANY,character}(serverInfo, name)
}
\arguments{
\item{serverInfo}{a WTSPS server object or URL}

\item{name}{An Algorithm name}
}
\description{
Returns an Algorithm Class in a WTSPS server URL queried by name
}
