% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-WTSCS.R
\docType{methods}
\name{getAlgorithms}
\alias{getAlgorithms}
\alias{getAlgorithms-generic}
\alias{getAlgorithms,WTSCS-method}
\title{Returns the WTSCS object's algorithms}
\usage{
getAlgorithms(object)

\S4method{getAlgorithms}{WTSCS}(object)
}
\arguments{
\item{object}{A WTSCS object}
}
\description{
Returns the WTSCS object's algorithms
}
