% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Process.R
\docType{methods}
\name{getCommand}
\alias{getCommand}
\alias{getCommand-generic}
\alias{getCommand,Process-method}
\title{Returns a Process status}
\usage{
getCommand(processInfo)

\S4method{getCommand}{Process}(processInfo)
}
\arguments{
\item{processInfo}{A Process object}
}
\description{
Returns a Process status
}
