% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Algorithm.R
\docType{methods}
\name{getDescription}
\alias{getDescription}
\alias{getDescription-generic}
\alias{getDescription,Algorithm-method}
\title{Returns algorithm description}
\usage{
getDescription(object)

\S4method{getDescription}{Algorithm}(object)
}
\arguments{
\item{object}{An Algorithm object}
}
\description{
Returns algorithm description
}
