% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Process.R
\docType{methods}
\name{getId}
\alias{getId}
\alias{getId-generic}
\alias{getId,Process-method}
\title{Returns a Process id}
\usage{
getId(object)

\S4method{getId}{Process}(object)
}
\arguments{
\item{object}{A Process object}
}
\description{
Returns a Process id
}
