% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Algorithm.R
\docType{methods}
\name{getInputParameters}
\alias{getInputParameters}
\alias{getInputParameters-generic}
\alias{getInputParameters,Algorithm-method}
\title{Returns Algorithm parameters}
\usage{
getInputParameters(object)

\S4method{getInputParameters}{Algorithm}(object)
}
\arguments{
\item{object}{An Algorithm object}
}
\description{
Returns Algorithm parameters
}
