% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Algorithm.R
\docType{methods}
\name{getName}
\alias{getName}
\alias{getName-generic}
\alias{getName,Algorithm-method}
\title{Returns an Algorithm name}
\usage{
getName(object)

\S4method{getName}{Algorithm}(object)
}
\arguments{
\item{object}{An Algorithm object}
}
\description{
Returns an Algorithm name
}
