% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Algorithm.R
\docType{methods}
\name{getOutput}
\alias{getOutput}
\alias{getOutput-generic}
\alias{getOutput,Algorithm-method}
\title{Returns algorithm output}
\usage{
getOutput(object)

\S4method{getOutput}{Algorithm}(object)
}
\arguments{
\item{object}{An Algorithm object}
}
\description{
Returns algorithm output
}
