% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Process.R
\docType{methods}
\name{getStatus}
\alias{getStatus}
\alias{getStatus-generic}
\alias{getStatus,Process-method}
\title{Returns a Process status}
\usage{
getStatus(processInfo)

\S4method{getStatus}{Process}(processInfo)
}
\arguments{
\item{processInfo}{A Process object}
}
\description{
Returns a Process status
}
