% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-WTSCS.R
\docType{methods}
\name{managePermissions}
\alias{managePermissions}
\alias{managePermissions-generic}
\alias{managePermissions,ANY-method}
\title{Executes a Process status in a WTSPS server URL}
\usage{
managePermissions(userInfo)

\S4method{managePermissions}{ANY}(userInfo)
}
\arguments{
\item{userInfo}{a User object or a login}
}
\description{
Executes a Process status in a WTSPS server URL
}
