% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Process.R
\docType{methods}
\name{parametrizeAlgorithm}
\alias{parametrizeAlgorithm}
\alias{parametrizeAlgorithm-generic}
\alias{parametrizeAlgorithm,ANY-method}
\title{Check Algorithm parameters of a running Process object}
\usage{
parametrizeAlgorithm(serverInfo, processInfo)

\S4method{parametrizeAlgorithm}{ANY}(serverInfo, processInfo)
}
\arguments{
\item{serverInfo}{A WTSPS object or URL}

\item{processInfo}{A Process object or command}
}
\description{
Check Algorithm parameters of a running Process object
}
