% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseJSON}
\alias{parseJSON}
\title{Parse a JSON response}
\usage{
parseJSON(HTTP_response)
}
\arguments{
\item{HTTP_response}{a HTTP request}
}
\description{
This function parses a HTTP reponse in case is JSON
}
