% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseJsonResponse}
\alias{parseJsonResponse}
\title{Parse a JSON response}
\usage{
parseJsonResponse(json_response)
}
\arguments{
\item{json_response}{a server response}
}
\description{
This function parses a JSON reponse
}
