% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Process.R, R/methods-Process.R
\docType{class}
\name{Process-class}
\alias{Process-class}
\alias{Process}
\title{Wrapper function Algorithm}
\usage{
Process(serverURL, id)
}
\arguments{
\item{serverURL}{a WTSPS server URL}

\item{id}{a Process identifier}
}
\description{
Wrapper function Algorithm
}
\section{Slots }{

\describe{
\item{\code{id}:}{Attribute of class \code{"character"}, an identifier of a Process.}
\item{\code{status}:}{Attribute of class \code{"character"}, a status of a Process.}
}
}

