% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sendHttpRequest}
\alias{sendHttpRequest}
\title{Send a HTTP request}
\usage{
sendHttpRequest(http_request)
}
\arguments{
\item{http_request}{a HTTP request}
}
\description{
This function submits a HTTP request to a URL
}
