% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-WTSCS.R
\docType{methods}
\name{statusProcess}
\alias{statusProcess}
\alias{statusProcess-generic}
\alias{statusProcess,ANY-method}
\title{Returns a Process status in a WTSCS server URL}
\usage{
statusProcess(serverInfo, processInfo)

\S4method{statusProcess}{ANY}(serverInfo, processInfo)
}
\arguments{
\item{serverInfo}{a WTSCS object or URL}

\item{processInfo}{a Process object}
}
\description{
Returns a Process status in a WTSCS server URL
}
