% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-User.R, R/methods-User.R
\docType{class}
\name{User-class}
\alias{User-class}
\alias{User}
\title{Wrapper function User}
\usage{
User(serverURL, login)
}
\arguments{
\item{serverURL}{a WTSPS server URL}

\item{login}{a Process identifier}
}
\description{
Wrapper function User
}
\section{Slots }{

\describe{
\item{\code{login}:}{Attribute of class \code{"character"}, login of the User.}
\item{\code{permissions}:}{Attribute of class \code{"character"}, permissions of the User.}
}
}

