% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-WTSCS.R, R/methods-WTSCS.R
\docType{class}
\name{WTSCS-class}
\alias{WTSCS-class}
\alias{WTSCS}
\title{Wrapper function WTSPS}
\usage{
WTSCS(serverURL)
}
\arguments{
\item{serverURL}{a WTSPS server URL}
}
\description{
Wrapper function WTSPS
}
\section{Slots }{

\describe{
\item{\code{serverURL}:}{Attribute of class \code{"character"}, URL of the server.}
\item{\code{algorithms}:}{Attribute of class \code{"character"}, algorithms of the server.}
}
}

