% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-WTSPS.R, R/methods-WTSPS.R
\docType{class}
\name{WTSPS-class}
\alias{WTSPS-class}
\alias{WTSPS}
\title{Wrapper function WTSPS}
\usage{
WTSPS(serverURL)
}
\arguments{
\item{serverURL}{a WTSPS server URL}
}
\description{
Wrapper function WTSPS
}
\section{Slots }{

\describe{
\item{\code{serverURL}:}{Attribute of class \code{"character"}, URL of the server.}
\item{\code{algorithms}:}{Attribute of class \code{"character"}, algorithms of the server.}
}
}

