% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-WTSPS.R
\docType{methods}
\name{getAlgorithms}
\alias{getAlgorithms}
\alias{getAlgorithms-generic}
\alias{getAlgorithms,WTSPS-method}
\title{Returns the WTSPS object's algorithms}
\usage{
getAlgorithms(serverInfo)

\S4method{getAlgorithms}{WTSPS}(serverInfo)
}
\arguments{
\item{serverInfo}{A WTSPS object}
}
\description{
Returns the WTSPS object's algorithms
}
