% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-WTSPS.R
\docType{methods}
\name{getAlgorithms}
\alias{getAlgorithms}
\alias{getAlgorithms-generic}
\alias{getAlgorithms,WTSPS-method}
\title{Returns the WTSPS object's algorithms}
\usage{
getAlgorithms(object)

\S4method{getAlgorithms}{WTSPS}(object)
}
\arguments{
\item{object}{A WTSPS object}
}
\description{
Returns the WTSPS object's algorithms
}
