% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-WTSCS.R
\docType{methods}
\name{getServerURL}
\alias{getServerURL}
\alias{getServerURL-generic}
\alias{getServerURL,WTSCS-method}
\title{Returns the WTSCS object's server URL}
\usage{
getServerURL(object)

\S4method{getServerURL}{WTSCS}(object)
}
\arguments{
\item{object}{A WTSCS object}
}
\description{
Returns the WTSCS object's server URL
}
