% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-WTSPS.R
\docType{methods}
\name{getServerURL}
\alias{getServerURL}
\alias{getServerURL-generic}
\alias{getServerURL,WTSPS-method}
\title{Returns the WTSPS object's server URL}
\usage{
getServerURL(serverInfo)

\S4method{getServerURL}{WTSPS}(serverInfo)
}
\arguments{
\item{serverInfo}{A WTSPS object}
}
\description{
Returns the WTSPS object's server URL
}
