% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Process.R
\docType{methods}
\name{getUuid}
\alias{getUuid}
\alias{getUuid-generic}
\alias{getUuid,Process-method}
\title{Returns a Process id}
\usage{
getUuid(object)

\S4method{getUuid}{Process}(object)
}
\arguments{
\item{object}{A Process object}
}
\description{
Returns a Process id
}
