% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-WTSPS.R
\docType{methods}
\name{listAlgorithms}
\alias{listAlgorithms}
\alias{listAlgorithms-generic}
\alias{listAlgorithms,ANY-method}
\title{Returns the WTSPS object's algorithms}
\usage{
listAlgorithms(serverInfo)

\S4method{listAlgorithms}{ANY}(serverInfo)
}
\arguments{
\item{serverInfo}{A WTSPS object or URL}
}
\description{
Returns the WTSPS object's algorithms
}
