% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Process.R, R/methods-Process.R
\docType{class}
\name{Process-class}
\alias{Process-class}
\alias{Process}
\title{Wrapper function Process}
\usage{
Process(serverInfo, processInfo, processStatus)
}
\arguments{
\item{serverInfo}{a WTSCS server URL}

\item{processInfo}{a Process identifier}

\item{processStatus}{a Process identifier}
}
\description{
Wrapper function Process
}
\section{Slots }{

\describe{
\item{\code{uuid}:}{Attribute of class \code{"character"}, an identifier of a Process.}
\item{\code{status}:}{Attribute of class \code{"character"}, a status of a Process.}
\item{\code{command}:}{Attribute of class \code{"character"}, a command of a Process.}
}
}

