% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-WTSCS.R
\docType{methods}
\name{runProcess}
\alias{runProcess}
\alias{runProcess-generic}
\alias{runProcess,ANY-method}
\title{Executes a Process command in a WTSCS server URL}
\usage{
runProcess(serverInfo, ...)

\S4method{runProcess}{ANY}(serverInfo, ...)
}
\arguments{
\item{serverInfo}{a WTSCS object or URL}

\item{...}{remaining Algorithm parameters}
}
\description{
Executes a Process command in a WTSCS server URL
}
