% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-WTSPS.R
\docType{methods}
\name{statusProcess}
\alias{statusProcess}
\alias{statusProcess-generic}
\alias{statusProcess,ANY-method}
\title{Returns a Process status in a WTSPS server URL}
\usage{
statusProcess(serverInfo, processInfo)

\S4method{statusProcess}{ANY}(serverInfo, processInfo)
}
\arguments{
\item{serverInfo}{a WTSPS object or URL}

\item{processInfo}{a Process object}
}
\description{
Returns a Process status in a WTSPS server URL
}
